requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
PyPDF2
pymupdf
langchain-community
langchain-huggingface
pypdf
sentence-transformers
fuzzywuzzy[speedup]
llama-index
ollama